+++ instrumentation/opentelemetry-instrumentation-pymongo/src/opentelemetry/instrumentation/pymongo/utils.py
@@ -12,9 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-COMMAND_TO_ATTRIBUTE_MAPPING = {
-    "insert": "documents",
-    "delete": "deletes",
-    "update": "updates",
-    "find": "filter",
+COMMAND_TO_ATTRIBUTES_MAPPING = {
+    "insert": ["documents"],
+    "delete": ["deletes"],
+    "update": ["updates"],
+    "find": ["filter"],
+    "findAndModify": ["query", "update"],
+    "aggregate": ["pipeline"],
+    "createIndexes": ["indexes"],
 }